XmlConfigTests - Changed to isolated LogFactory
